+++ src/features/guidUtilsProvider.ts
@@ -2,6 +2,7 @@ import path = require('path');
 import * as vscode from 'vscode';
 import * as fs from 'fs';
 import { AssetsTocProvider } from '../other/assetsTocProvider';
+import * as xmldoc from 'xmldoc';
 
 const _TAGS_TO_COMPLETE: { [index: string]: string[] } = {
   /* eslint-disable @typescript-eslint/naming-convention */
@@ -10,6 +11,7 @@ const _TAGS_TO_COMPLETE: { [index: string]: string[] } = {
   'ItemLink': [ 'GuildhouseItem', 'HarborOfficeItem', 'TownhallItem', 'CultureItem', 'VehicleItem' ],
   'Good': [ 'Product' ], // TODO and items?
   'GUID': [ '*' ],
+  'Building': [ '*' ],
   'ProvidedNeed': [ 'Product' ],
   'SubstituteNeed': [ 'Product' ],
   /* eslint-enable @typescript-eslint/naming-convention */
@@ -23,11 +25,12 @@ interface IAsset {
 }
 
 export function resolveGUID(guid: string) {
-  const vanilla = _vanillaAssets || {};
-
   let entry = undefined;
-  if (vanilla) {
-    entry = vanilla[guid];
+  if (_vanillaAssets) {
+    entry = _vanillaAssets[guid];
+  }
+  if (_customAssets && !entry) {
+    entry = _customAssets[guid];
   }
   return entry;
 }
@@ -191,34 +194,44 @@ function getValueAt(line: string, position: number) {
   };
 }
 
-const _completionItems: { [ index: string]: vscode.CompletionItem[] } = {};
+function _updateCompletionItems(assets: { [index: string]: IAsset }) {
+  const completionItems: { [ index: string]: vscode.CompletionItem[] } = {};
+  
+  // prepare CompletionItem lists
+  for (let tag of Object.keys(_TAGS_TO_COMPLETE)) {
+    completionItems[tag] = [];
+  }
+
+  // fill CompletionItem lists for the different kind of tags
+  for (let guid of Object.keys(assets)) {
+    const asset = assets[guid];
+    const item = new vscode.CompletionItem({
+      label: `${asset.english||asset.name}`,
+      description: `${asset.template}: ${guid} (${asset.name})`
+    }, vscode.CompletionItemKind.Snippet);
+    item.insertText = guid;
+
+    for (let tag of Object.keys(_TAGS_TO_COMPLETE)) {
+      if (_TAGS_TO_COMPLETE[tag].indexOf('*') !== -1 || asset.template && _TAGS_TO_COMPLETE[tag].indexOf(asset.template) !== -1) {
+        completionItems[tag].push(item);
+      }
+    }
+  }
+
+  return completionItems;
+}
+
+let _completionItems: { [ index: string]: vscode.CompletionItem[] } = {};
+let _vanillaCompletionItems: { [ index: string]: vscode.CompletionItem[] } = {};
+let _customCompletionItems: { [ index: string]: vscode.CompletionItem[] } | undefined = undefined;
 let _vanillaAssets: { [index: string]: IAsset } | undefined = undefined;
 async function loadVanillaAssets(context: vscode.ExtensionContext) {
   if (!_vanillaAssets) {
     const assetPath = context.asAbsolutePath('./generated/assets.json');
     _vanillaAssets = JSON.parse(fs.readFileSync(assetPath, { encoding: 'utf8' }));
-
     if (_vanillaAssets) {
-      // prepare CompletionItem lists
-      for (let tag of Object.keys(_TAGS_TO_COMPLETE)) {
-        _completionItems[tag] = [];
-      }
-
-      // fill CompletionItem lists for the different kind of tags
-      for (let guid of Object.keys(_vanillaAssets)) {
-        const asset = _vanillaAssets[guid];
-        const item = new vscode.CompletionItem({ label: `${asset.english}`, description: `${asset.template}: ${guid} (${asset.name})` }, vscode.CompletionItemKind.Snippet);
-        item.insertText = guid;
-
-        for (let tag of Object.keys(_TAGS_TO_COMPLETE)) {
-          for (let template of _TAGS_TO_COMPLETE[tag]) {
-            if (template === '*' || template === asset.template) {
-              _completionItems[tag].push(item);
-              break;
-            }
-          }
-        }
-      }
+      _vanillaCompletionItems = _updateCompletionItems(_vanillaAssets);
+      _completionItems = _vanillaCompletionItems;
     }
   }
 
@@ -256,10 +269,81 @@ async function loadKeywordHelp(context: vscode.ExtensionContext) {
   return _keywordHelp;
 }
 
+let _customAssets: { [index: string]: IAsset } | undefined = undefined;
+export function refreshCustomAssets(document: vscode.TextDocument | undefined): void {
+  if (!document) {
+    _customAssets = undefined;
+    _customCompletionItems = undefined;
+    return;
+  }
+
+  _customAssets = {};
+
+  const relevantNodes = new Set<string>(['ModOps', 'ModOp', 'Asset', 'Values', 'Standard', 'GUID']);
+
+  const xmlContent = new xmldoc.XmlDocument(document.getText());
+  const nodeStack: { history: xmldoc.XmlElement[], element: xmldoc.XmlNode }[] = [{ history: [], element: xmlContent }];
+  while (nodeStack.length > 0) {
+    const top = nodeStack.pop();
+    if (top?.element.type === 'element' && relevantNodes.has(top.element.name)) {
+      const name = top.element.name;
+      if (top.element.name === 'GUID') {
+        const guid = top.element.val;
+        const parent = top.history.length >= 2 ? top.history[top.history.length - 2] : undefined;
+        const name = parent?.valueWithPath('Name');
+
+        if (name) {
+          _customAssets[guid] = {
+            guid,
+            name,
+            template: top.history.length >= 4 ? top.history[top.history.length - 4].valueWithPath('Template') : undefined
+          };
+        }
+      }
+
+      const children = (top.element.children ? top.element.children.filter((e) => e.type === 'element') : []).map((e) => (
+        { history: [...top.history, e as xmldoc.XmlElement], element: e }
+      ));
+      if (children.length > 0) {
+        // has tag children
+        nodeStack.push(...children.reverse());
+      }
+    }
+    else {
+      // ignore
+    }
+  }
+
+  _customCompletionItems = _updateCompletionItems(_customAssets);
+}
+
+function subscribeToDocumentChanges(context: vscode.ExtensionContext): void {
+  if (vscode.window.activeTextEditor) {
+    refreshCustomAssets(vscode.window.activeTextEditor.document);
+  }
+  context.subscriptions.push(
+    vscode.window.onDidChangeActiveTextEditor(editor => {
+      if (editor) {
+        refreshCustomAssets(editor.document);
+      }
+    })
+  );
+
+  context.subscriptions.push(
+    vscode.workspace.onDidChangeTextDocument(e => refreshCustomAssets(e.document))
+  );
+
+  context.subscriptions.push(
+    vscode.workspace.onDidCloseTextDocument(doc => refreshCustomAssets(undefined))
+  );
+
+}
+
 export function registerGuidUtilsProvider(context: vscode.ExtensionContext): vscode.Disposable[] {
   loadVanillaAssets(context);
   loadGuidRanges(context);
   loadKeywordHelp(context);
+  subscribeToDocumentChanges(context);
 
 	return [
     vscode.Disposable.from(vscode.languages.registerHoverProvider({ language: 'xml', pattern: '**/assets.xml' }, { provideHover })), 
@@ -272,7 +356,14 @@ function provideCompletionItems(document: vscode.TextDocument, position: vscode.
   if (!keyword) {
     return undefined;
   }
-  // TODO scan open file for GUIDs
+
+  if (_customCompletionItems) {
+    const customItems = _customCompletionItems[keyword.name];
+    if (customItems) {
+      return [ ..._vanillaCompletionItems[keyword.name], ...customItems ];
+    }
+  }
+
   return _completionItems[keyword.name];
 }
 
@@ -300,7 +391,12 @@ function provideHover(document: vscode.TextDocument, position: vscode.Position, 
       const namedGuid = resolveGUID(guid);
       let name = [ ];
       if (namedGuid) {
-        name = [ `${namedGuid.template}: ${namedGuid.english} (${namedGuid.name})` ];
+        if (namedGuid.english) {
+          name = [ `${namedGuid.template}: ${namedGuid.english} (${namedGuid.name})` ];
+        }
+        else {
+          name = [ `${namedGuid.template}: ${namedGuid.name}` ];
+        }
       }
       else {
         name = [ `GUID ${guid} not found.` ];